PWA: Update check rework per feedback
