+++ modern/src/App.jsx
@@ -8,7 +8,7 @@ import SocketController from './SocketController';
 import CachingController from './CachingController';
 import { useEffectAsync } from './reactHelper';
 import { sessionActions } from './store';
-import ReloadPrompt from './ReloadPrompt';
+import UpdateCheckPrompt from './UpdateCheckPrompt';
 
 const useStyles = makeStyles(() => ({
   page: {
@@ -49,7 +49,7 @@ const App = () => {
     <>
       <SocketController />
       <CachingController />
-      <ReloadPrompt />
+      <UpdateCheckPrompt />
       <div className={classes.page}>
         <Outlet />
       </div>
