+++ modern/src/resources/l10n/en.json
@@ -99,8 +99,6 @@
     "sharedImport": "Import",
     "sharedColumns": "Columns",
     "sharedDropzoneText": "Drag and drop a file here or click",
-    "sharedReload": "Reload",
-    "updateAvailable": "New update available, click on the reload button to update.",
     "calendarSimple": "Simple",
     "calendarRecurrence": "Recurrence",
     "calendarOnce": "Once",
@@ -229,6 +227,8 @@
     "settingsTotpEnable": "Enable One-time Password",
     "settingsTotpForce": "Force One-time Password",
     "settingsServiceWorkerUpdateInterval": "ServiceWorker Update Interval",
+    "settingsReload": "Reload",
+    "settingsUpdateAvailable": "New update available, click on the reload button to update.",
     "reportTitle": "Reports",
     "reportScheduled": "Scheduled Reports",
     "reportDevice": "Device",
